requirements.txt
# Global Python dependencies
faker
boto3
pyyaml
pandas
pyarrow